+++ sites/jeromefitzgerald.com/src/components/icon/icon.custom.tsx
@@ -1,0 +1,154 @@
+"use client";
+
+import { AccessibleIcon } from "@radix-ui/react-accessible-icon";
+
+import type { IconProps } from "./icon.types";
+
+const AmazonLogoIcon = ({ label, ...props }: IconProps) => (
+  <AccessibleIcon
+    label={
+      label ||
+      "An icon representing the logo of Amazon. Which is a sans serif letter -a, with a curved arrow underline pointing to the right that mimics a smirk or smiley face."
+    }
+  >
+    <svg
+      className="fill:bg-black dark:fill:bg-white"
+      height="17px"
+      role="img"
+      version="1.1"
+      viewBox="3 1 252 260"
+      width="17px"
+      xmlns="http://www.w3.org/2000/svg"
+      xmlnsXlink="http://www.w3.org/1999/xlink"
+      {...props}
+    >
+      <g fill="none" fill-rule="evenodd">
+        <path
+          d="m221.503 210.324c-105.235 50.083-170.545 8.18-212.352-17.271-2.587-1.604-6.984.375-3.169 4.757 13.928 16.888 59.573 57.593 119.153 57.593 59.621 0 95.09-32.532 99.527-38.207 4.407-5.627 1.294-8.731-3.16-6.872zm29.555-16.322c-2.826-3.68-17.184-4.366-26.22-3.256-9.05 1.078-22.634 6.609-21.453 9.93.606 1.244 1.843.686 8.06.127 6.234-.622 23.698-2.826 27.337 1.931 3.656 4.79-5.57 27.608-7.255 31.288-1.628 3.68.622 4.629 3.68 2.178 3.016-2.45 8.476-8.795 12.14-17.774 3.639-9.028 5.858-21.622 3.71-24.424z"
+          fill="currentColor"
+          fill-rule="nonzero"
+        ></path>
+        <path
+          d="m150.744 108.13c0 13.141.332 24.1-6.31 35.77-5.361 9.489-13.853 15.324-23.341 15.324-12.952 0-20.495-9.868-20.495-24.432 0-28.75 25.76-33.968 50.146-33.968zm34.015 82.216c-2.23 1.992-5.456 2.135-7.97.806-11.196-9.298-13.189-13.615-19.356-22.487-18.502 18.882-31.596 24.527-55.601 24.527-28.37 0-50.478-17.506-50.478-52.565 0-27.373 14.85-46.018 35.96-55.126 18.313-8.066 43.884-9.489 63.43-11.718v-4.365c0-8.018.616-17.506-4.08-24.432-4.128-6.215-12.003-8.777-18.93-8.777-12.856 0-24.337 6.594-27.136 20.257-.57 3.037-2.799 6.026-5.835 6.168l-32.735-3.51c-2.751-.618-5.787-2.847-5.028-7.07 7.543-39.66 43.36-51.616 75.43-51.616 16.415 0 37.858 4.365 50.81 16.795 16.415 15.323 14.849 35.77 14.849 58.02v52.565c0 15.798 6.547 22.724 12.714 31.264 2.182 3.036 2.657 6.69-.095 8.966-6.879 5.74-19.119 16.415-25.855 22.393l-.095-.095"
+          fill="currentColor"
+        ></path>
+        <path
+          d="m221.503 210.324c-105.235 50.083-170.545 8.18-212.352-17.271-2.587-1.604-6.984.375-3.169 4.757 13.928 16.888 59.573 57.593 119.153 57.593 59.621 0 95.09-32.532 99.527-38.207 4.407-5.627 1.294-8.731-3.16-6.872zm29.555-16.322c-2.826-3.68-17.184-4.366-26.22-3.256-9.05 1.078-22.634 6.609-21.453 9.93.606 1.244 1.843.686 8.06.127 6.234-.622 23.698-2.826 27.337 1.931 3.656 4.79-5.57 27.608-7.255 31.288-1.628 3.68.622 4.629 3.68 2.178 3.016-2.45 8.476-8.795 12.14-17.774 3.639-9.028 5.858-21.622 3.71-24.424z"
+          fill="currentColor"
+          fill-rule="nonzero"
+        ></path>
+        <path
+          d="m150.744 108.13c0 13.141.332 24.1-6.31 35.77-5.361 9.489-13.853 15.324-23.341 15.324-12.952 0-20.495-9.868-20.495-24.432 0-28.75 25.76-33.968 50.146-33.968zm34.015 82.216c-2.23 1.992-5.456 2.135-7.97.806-11.196-9.298-13.189-13.615-19.356-22.487-18.502 18.882-31.596 24.527-55.601 24.527-28.37 0-50.478-17.506-50.478-52.565 0-27.373 14.85-46.018 35.96-55.126 18.313-8.066 43.884-9.489 63.43-11.718v-4.365c0-8.018.616-17.506-4.08-24.432-4.128-6.215-12.003-8.777-18.93-8.777-12.856 0-24.337 6.594-27.136 20.257-.57 3.037-2.799 6.026-5.835 6.168l-32.735-3.51c-2.751-.618-5.787-2.847-5.028-7.07 7.543-39.66 43.36-51.616 75.43-51.616 16.415 0 37.858 4.365 50.81 16.795 16.415 15.323 14.849 35.77 14.849 58.02v52.565c0 15.798 6.547 22.724 12.714 31.264 2.182 3.036 2.657 6.69-.095 8.966-6.879 5.74-19.119 16.415-25.855 22.393l-.095-.095"
+          fill="currentColor"
+        ></path>
+      </g>
+    </svg>
+  </AccessibleIcon>
+);
+
+const AppleLogoIcon = ({ label, ...props }: IconProps) => (
+  <AccessibleIcon label={label || "An icon representing the logo of Apple. Which is an apple."}>
+    <svg
+      className="fill:bg-black dark:fill:bg-white"
+      height="17px"
+      role="img"
+      version="1.1"
+      viewBox="0 0 814 1000"
+      width="17px"
+      xmlns="http://www.w3.org/2000/svg"
+      xmlnsXlink="http://www.w3.org/1999/xlink"
+      {...props}
+    >
+      <path
+        d="M788.1 340.9c-5.8 4.5-108.2 62.2-108.2 190.5 0 148.4 130.3 200.9 134.2 202.2-.6 3.2-20.7 71.9-68.7 141.9-42.8 61.6-87.5 123.1-155.5 123.1s-85.5-39.5-164-39.5c-76.5 0-103.7 40.8-165.9 40.8s-105.6-57-155.5-127C46.7 790.7 0 663 0 541.8c0-194.4 126.4-297.5 250.8-297.5 66.1 0 121.2 43.4 162.7 43.4 39.5 0 101.1-46 176.3-46 28.5 0 130.9 2.6 198.3 99.2zm-234-181.5c31.1-36.9 53.1-88.1 53.1-139.3 0-7.1-.6-14.3-1.9-20.1-50.6 1.9-110.8 33.7-147.1 75.8-28.5 32.4-55.1 83.6-55.1 135.5 0 7.8 1.3 15.6 1.9 18.1 3.2.6 8.4 1.3 13.6 1.3 45.4 0 102.5-30.4 135.5-71.3z"
+        fill="currentColor"
+        id="apple-icon"
+      ></path>
+    </svg>
+  </AccessibleIcon>
+);
+
+const BlueskyLogoIcon = ({ label, ...props }: IconProps) => (
+  <AccessibleIcon
+    label={
+      label || "An icon representing the logo of Bluesky. Which is a silhouette of a butterfly."
+    }
+  >
+    <svg
+      className="fill:bg-black dark:fill:bg-white"
+      height="17px"
+      role="img"
+      version="1.1"
+      viewBox="0 0 568 501"
+      width="17px"
+      xmlns="http://www.w3.org/2000/svg"
+      xmlnsXlink="http://www.w3.org/1999/xlink"
+      {...props}
+    >
+      <path
+        d="M123.121 33.664C188.241 82.553 258.281 181.68 284 234.873c25.719-53.192 95.759-152.32 160.879-201.21C491.866-1.611 568-28.906 568 57.947c0 17.346-9.945 145.713-15.778 166.555-20.275 72.453-94.155 90.933-159.875 79.748C507.222 323.8 536.444 388.56 473.333 453.32c-119.86 122.992-172.272-30.859-185.702-70.281-2.462-7.227-3.614-10.608-3.631-7.733-.017-2.875-1.169.506-3.631 7.733-13.43 39.422-65.842 193.273-185.702 70.281-63.111-64.76-33.89-129.52 80.986-149.071-65.72 11.185-139.6-7.295-159.875-79.748C9.945 203.659 0 75.291 0 57.946 0-28.906 76.135-1.612 123.121 33.664Z"
+        fill="currentColor"
+        id="bluesky-icon"
+      ></path>
+    </svg>
+  </AccessibleIcon>
+);
+
+const GooglePodcastsLogoIcon = ({ label, ...props }: IconProps) => (
+  <AccessibleIcon label={label || "An icon representing the logo for Google Podcasts."}>
+    <svg
+      className="fill:bg-black dark:fill:bg-white"
+      height="17px"
+      role="img"
+      version="1.1"
+      viewBox="0 0 400 400"
+      width="17px"
+      xmlns="http://www.w3.org/2000/svg"
+      xmlnsXlink="http://www.w3.org/1999/xlink"
+      {...props}
+    >
+      <g stroke-linecap="round" stroke-width="48.5">
+        <path d="m24.26 186.32v27.351" stroke="currentColor"></path>
+        <path d="m375.74 186.32v27.351" stroke="currentColor"></path>
+        <path d="m112.13 104.27v89.46m0 74.645v27.351" stroke="currentColor"></path>
+        <path d="m287.87 206.26v89.46m0-191.46v27.351" stroke="currentColor"></path>
+        <path d="m200 127.06v145.87m0-248.68v27.351m0 296.78v27.351" stroke="currentColor"></path>
+      </g>
+    </svg>
+  </AccessibleIcon>
+);
+
+const SpotifyLogoIcon = ({ label, ...props }: IconProps) => (
+  <AccessibleIcon
+    label={
+      label ||
+      "An icon representing the logo of Spotify. A circle with what looks like three sound waves slightly off centered to the right and pointed upward. The waves start small and get larger, and encompass about 90% of the circle with some space around them. The slightly off centered icon rankles a lot of people."
+    }
+  >
+    <svg
+      className="fill:bg-black dark:fill:bg-white"
+      height="17px"
+      role="img"
+      version="1.1"
+      viewBox="0 0 17 17"
+      width="17px"
+      xmlns="http://www.w3.org/2000/svg"
+      xmlnsXlink="http://www.w3.org/1999/xlink"
+      {...props}
+    >
+      <defs></defs>
+      <g fill="none" fillRule="evenodd" stroke="none" strokeWidth="1">
+        <g fill="currentColor" transform="translate(-303.000000, -5776.000000)">
+          <g id="Group-12" transform="translate(130.000000, 5758.000000)">
+            <path
+              d="M186.918938,25.6499188 C186.791639,25.6499188 186.662513,25.6199721 186.542015,25.5566271 C181.694705,23.0107486 176.716441,24.5350881 176.666699,24.5507213 C176.238816,24.685837 175.782712,24.4483949 175.647799,24.0208164 C175.512785,23.5930349 175.750024,23.1369309 176.177704,23.0019168 C176.407126,22.9295371 181.857432,21.2615547 187.297282,24.1186763 C187.694304,24.3272883 187.847185,24.8182117 187.638573,25.2152343 C187.493408,25.4918609 187.210995,25.6499188 186.918938,25.6499188 M186.447099,27.9201858 C186.325079,28.1324523 186.102864,28.2512241 185.874457,28.2512241 C185.762892,28.2512241 185.649805,28.2229016 185.546159,28.1633127 C181.341333,25.7451393 177.105241,27.1529403 177.062909,27.1673554 C176.718166,27.2855181 176.342969,27.1019801 176.224705,26.7572374 C176.106542,26.4124946 176.290182,26.0371961 176.634925,25.9190333 C176.830746,25.8519324 181.485585,24.3058687 186.204074,27.0192459 C186.519986,27.2009566 186.62881,27.6042732 186.447099,27.9201858 M185.391553,30.272984 C185.28679,30.4471827 185.102034,30.5436213 184.912507,30.5436213 C184.814647,30.5436213 184.715366,30.5178367 184.625322,30.463628 C181.277988,28.4501923 177.310198,29.614765 177.270506,29.6267437 C176.975708,29.7157717 176.663856,29.5493897 176.57432,29.2546935 C176.484785,28.9598959 176.650456,28.6484498 176.945152,28.5585081 C177.126152,28.5032843 181.418992,27.231817 185.200909,29.5067537 C185.465151,29.6657252 185.550524,30.0087422 185.391553,30.272984 M181.5,18 C176.80557,18 173,21.8055702 173,26.5001015 C173,31.1944298 176.80557,35 181.5,35 C186.19443,35 190,31.1944298 190,26.5001015 C190,21.8055702 186.19443,18 181.5,18"
+              id="spotify-icon"
+            ></path>
+          </g>
+        </g>
+      </g>
+    </svg>
+  </AccessibleIcon>
+);
+
+export { AmazonLogoIcon, AppleLogoIcon, BlueskyLogoIcon, GooglePodcastsLogoIcon, SpotifyLogoIcon };
